LLM-Finetuner-main/requirements.txt
Flask==3.1.0
requests==2.32.3
flask_sqlalchemy==3.1.1
flask_migrate==4.0.7
flask_cors==5.0.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
python-dotenv==1.0.1
runpod==1.7.7
psycopg2-binary
pymilvus==2.3.4
numpy>=1.24.0
aiofiles>=23.0.0
openai>=1.0.0
llama-index>=0.12.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-readers-file>=0.4.0
llama-index-readers-web>=0.3.0
selenium>=4.0.0
webdriver-manager>=4.0.0
pdf2image>=1.16.0
beautifulsoup4>=4.12.0
python-pptx>=0.6.0
Pillow>=10.0.0